Breaks tests into smaller classes
